TPM fix after LV update
Fixed up TPM after adding Omega controller to LabView. The Omega output
goes to AI0, and replaces the mass flow output that was corrected by the
knob on the MKS 246. The omega output adds to the DAQ output file in the
temperature block, and the redundant entry for AI0 was removed. This
required some handling of column offsets. -- Alexis